requirements.txt
azure-ai-inference
python-dotenv
langchain-community
rank_bm25
pandas
requests
